Stop truncating every character above U+FFFF
juce_wchar is 32 bits. The generator mapped it to uint32 by name, but libclang
resolves the alias to wchar_t before that mapping is consulted, and wchar_t was
mapped to uint16. Everything that returns a character went through the wrong
one.

It did not fail. String's operator[] returned 62976 for U+1F600, which is
0x1F600 cut to sixteen bits, and getAndAdvance did the same. A caller reading
anything outside the basic plane got a different character back and no
indication of it.

wchar_t is 32 bits on macOS and Linux, which are the platforms this binds.
JUCE only spells a type as wchar_t where wchar_t is 16-bit - Windows - and
CharPointer_UTF16 uses int16 there instead, so nothing on these platforms wants
the old mapping.

112 bindings change type. A test reads U+1F600 back through operator[] and
through the character pointer.

Found while writing coverage for CharacterFunctions, which is also where the
next defect showed up: isDigit and isWhitespace are each declared twice, once
for char and once for juce_wchar, and an argument from Nim converts to both, so
C++ cannot pick either. Those two are left out of the test with that written
down; the fix is a separate change.

+++ sources/june/juce_core_lifting.nim
@@ -80,7 +80,7 @@ include juce_core_subclasses
 # String and MemoryBlock both expose a C++ iterator, which has no Nim spelling,
 # and neither had a Nim one to loop with instead.
 
-iterator items*(this: String): uint16 =
+iterator items*(this: String): uint32 =
     for index in 0 ..< this.length():
         yield this[index]
